+++ json_annotation/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.7.1-dev
+
+- Update `JsonKey` documentation to align with new features in 
+  `package:json_serializable`.
+
 ## 4.7.0
 
 - Added `JsonEnum.valueField` which allows specifying a field in an